feat: Reels removed refactor: remove unused Reels components and update import paths for validation schemas
