--- src/components/forms/ChatForm.tsx
@@ -1,3 +1,0 @@
-const ChatForm = () => {}
-
-export default ChatForm
