+++ src/components/forms/ProfileForm.tsx
@@ -13,7 +13,7 @@ import { Input } from '@/components/ui/input'
 import { Skeleton } from '@/components/ui/skeleton'
 import { Textarea } from '@/components/ui/textarea'
 import { useToast } from '@/components/ui/use-toast'
-import { ProfileValidationSchema } from '@/lib/validations/profile.validation.schema'
+import { ProfileValidationSchema } from '@/lib/validations/schemas/profile.validation.schema'
 import { useUpdateUser } from '@/states/query/hooks/mutations'
 import { useUser } from '@/states/query/hooks/queries'
 import { zodResolver } from '@hookform/resolvers/zod'
