+++ src/components/shared/chats/ChatBubble.tsx
@@ -31,7 +31,7 @@ import { Form, FormControl, FormField, FormItem } from '@/components/ui/form'
 import { Textarea } from '@/components/ui/textarea'
 import { useToast } from '@/components/ui/use-toast'
 import { cn, multiFormatDateString } from '@/lib/utils'
-import { EditionMessageValidationSchema } from '@/lib/validations/edition-message.validation.schema'
+import { EditionMessageValidationSchema } from '@/lib/validations/schemas/edition-message.validation.schema'
 import {
   useDeleteMessage,
   useEditMessage
