+++ src/components/shared/saves/SavedCollections.tsx
@@ -1,4 +1,4 @@
-import { NotFound } from '@/layouts/_root/pages'
+import NotFound from '@/layouts/_root/pages/NotFound'
 
 const SavedCollections = () => <NotFound />
 export default SavedCollections
