--- src/components/shared/saves/SavedReels.tsx
@@ -1,6 +1,0 @@
-import { NotFound } from '@/layouts/_root/pages'
-
-const SavedReels = () => {
-  return <NotFound />
-}
-export default SavedReels
