--- src/layouts/_root/pages/Reels.tsx
@@ -1,9 +1,0 @@
-import { NotFound } from '@/_root/pages'
-
-const Reels = () => {
-  return (
-    <NotFound />
-  )
-}
-
-export default Reels
