--- src/layouts/_root/pages/index.ts
@@ -1,13 +1,0 @@
-export { default as Chats } from '@/_root/pages/Chats'
-export { default as CreatePost } from '@/_root/pages/CreatePost'
-export { default as EditPost } from '@/_root/pages/EditPost'
-export { default as Explore } from '@/_root/pages/Explore'
-export { default as Home } from '@/_root/pages/Home'
-export { default as LikedPosts } from '@/_root/pages/LikedPosts'
-export { default as NotFound } from '@/_root/pages/NotFound'
-export { default as People } from '@/_root/pages/People'
-export { default as PostDetails } from '@/_root/pages/PostDetails'
-export { default as Profile } from '@/_root/pages/Profile'
-export { default as Reels } from '@/_root/pages/Reels'
-export { default as Saved } from '@/_root/pages/Saved'
-export { default as UpdateProfile } from '@/_root/pages/UpdateProfile'
