+++ src/main.tsx
@@ -1,3 +1,4 @@
+import App from '@/App'
 import LazyPages from '@/components/shared/app/LazyPages'
 import AccountProvider from '@/states/account/providers/AccountProvider'
 import QueryProvider from '@/states/query/providers/QueryProvider'
@@ -5,7 +6,6 @@ import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
-import App from './App'
 
 const ROOT = document.getElementById('root')
 if (ROOT == null) throw new Error('Root element not found')
